Use size_t for lengths in AT driver

+++ libgammu/protocol/at/at.h
@@ -11,8 +11,8 @@ GSM_Error AT_Initialise(GSM_StateMachine *s);
 typedef struct {
 	GSM_Protocol_Message 	Msg;
 	gboolean 			wascrlf;
-	int			LineStart,LineEnd;
-	int			SpecialAnswerLines,SpecialAnswerStart;
+	size_t			LineStart, LineEnd;
+	size_t			SpecialAnswerLines, SpecialAnswerStart;
 
 	gboolean			EditMode;	/* wait for modem answer or not */
 	gboolean			FastWrite;
